Delete ports in parallel when tearing down env
For the sake of simplicity, we delete all of the ports on the
provisioning and public networks before deleting the heat stack.
This is to ensure that the ports we attached to the undercloud vm
are removed before heat tries to delete the subnets.  In some jobs
this can mean we delete 12 ports in series.  Starting the deletes
in parallel should reduce the overall time it takes to delete the
ports by a fair amount since the controller is capable of working
on more than one request at a time.

Note that we're doing something similar for log collection in
If65e0e460467185a5695cf214272ef35e41c3f03 and it seems to be
working well there.

Change-Id: Iec46b14a5531b287ed70b3dffb7e42cb6bb855c9