Set DIB_GRUB_TIMEOUT to 0 for ci images
In ci there's no reason to wait for a grub prompt when booting
instances, so let's make use of the new config knob added in
I556b3f48eff1b67ee8c4b9b64f749af95100fb99 to set it to 0.

Note that this is ci-specific.  In real deployments there may be
reasons to want access to the grub prompt, so I don't think we
should shut it off by default in the image build.

Change-Id: Ia741c07239b4f0c5c1bfbb9df4ae46c94a568c43