Stop running if zuul job goes away
At times we are ending up with significantly more testenvs running
than we have zuul jobs.  The current theory is that this happens
when a zuul job is removed before completion, possibly when a new
patch set is pushed or a change is abandoned.  Because our test
jobs are unaware of zuul they continue to run even though infra
thinks the job has been stopped.  This can cause us to exceed the
quota for the nodepool user.

This patch attempts to deal with the problem by periodically
querying the zuul status endpoint and looking for the ZUUL_UUID
value in the output.  If it's not found the job will be killed.

Change-Id: I31a8157366a6ec91d8a0befeadc51eca8d8225e1
Closes-Bug: 1655489