OOQ/OVB: Add configurations for tempest job
This change is about replacing current tempest job with a one using
quickstart, proposed in https://review.openstack.org/420647.
We specify a new name and job type containing "oooq" so quickstart
script will be activated, and inside this script we're checking if the
TOCI_JOBTYPE contains the pattern to trigger tempest.
We are changing the test to apply on nonha deployments instead of HA,
and setting the ansible variables accordingly.
The configuration tested locally for a successful full tempest run use
the minimal_pacemaker quickstart configuration, but overriding the
variables that activate the ping test and set the variables that
activate tempest test. Cinder-backup service is not enable by default
and we are enabling to make some tests pass.i
We're going to run the full list of tests, with some exceptions added to
the skip list.

Change-Id: I3084e020e6997efd804306b3350dc9752cda549e