Work around ceph problems on mitaka
There is a conflict between the base and ceph repos that we use on
the mitaka branch.  One solution is to use a repo priority to make
the ceph repo take precedence.  Until this problem is fixed by the
ceph people, let's do the priority workaround ourselves to get the
mitaka jobs passing again.

Change-Id: I6bd245852ac621272ffdaa879214b3f88590a489
Partial-Bug: 1654611