oooq-multinode: playbook to pass missing env vars and check overcloud deployment
Presently the tripleo-ci-centos-7-nonha-multinode-oooq-nv check is
failing because the TRIPLEO_ROOT environment variable is not being
passed to the overcloud deploy. This passed because the overcloud-deploy
ansible role ignores the errors. This will be corrected in another
commit, but here we should pass the TRIPLEO_ROOT environment variable.

Additionally a new task should be added to the multinode playbook that
fails the run if the overcloud deployment result is failure.

Change-Id: I2c2ea05bc827de534b9f0e43b5ebec1e3696c7d5