Use http delorean urls instead of https
We can't cache https urls in squid which hurts our job runtimes and
is probably driving excessive amounts of external traffic from the
cloud.  For ci purposes, let's use http instead.

Unfortunately, this doesn't entirely solve the problem.  The
delorean-deps http repo redirects to https.  That's something we'll
need to discuss with the people in charge of the repo, but in the
meantime we can start pointing at the http url and if/when we get
that working on the repo side we'll start to benefit from caching
of that repo too.

Change-Id: Icb88b7111f09ed302ace0800f4f6e2a898e0b450