Quickstart transition: replace toci_gate_test.sh
This change modifies the way CI jobs are handled.
Current toci_gate_test.sh is renamed to toci_gate_test-orig.sh, and its
content will not be changed. toci_gate_test-oooq.sh is added, with
toci_gate_test.sh modified to be a link to it.
This new script will drive jobs to be launched using quickstart, it will
work initially as additional layer with backwards compatibility to old
script that will run jobs with tripleo.sh.
The old set of playbooks specific to toci is copied (not moved) from the
experimental scripts, and the toci-quickstart dir, containing
configurations for quickstart specific to tripleo ci is added.
Script toci_quickstart.sh will be the new single runner for every job
which configuration is assembled through toci_gate_test-oooq.sh script
To transition a job to quickstart, the only thing needed is to change
the job name to contain "featureset" keyword.
Make sure that the correspondent featureset file and nodes configuration
exist.

Change-Id: Ieb181db13e63e046bffc2c65120c2c24177488c5