Add gate-tripleo-ci-centos-7-ovb-updates into promotion pipeline
When Ocata will be released, we want to add
gate-tripleo-ci-centos-7-ovb-updates in the promotion pipeline to be
able to detect issues related to networking & IPv6 when testing new
versions of OpenStack.

We have seen some cases where gate-tripleo-ci-centos-7-ovb-updates broke
because of some changes in OpenStack. Let's add it into the pipeline to
avoid this situation again.

Change-Id: I17dbd0028451d3441274bcd3eb1d247eb29d1fba