Add converge update to major upgrade test
Discussion on IRC has made clear that for real deployments we can't
avoid the converge step, because we do still need a way to pin/unpin
the RPC (even when using "auto" as the version), and also potentially
some other nova specific steps will be required during the converge.

So align the CI test with how real deployments will be upgraded, even
though this will make the job take a little longer, otherwise the test
isn't representative of the expected upgrade workflow.

Depends-On: If5016b910931364a621b280465420d0bf2617895
Change-Id: Ifcdb048a1081b16a8a33816be98ddbbe07c27856