Don't use delorean-ci repo for upgrade deployment
The initial deploy uses UPGRADE_RELEASE branch, which is the previous
release, not the branch under test.  This means, because Depends-On
and ZUUL_REFS aren't branch-aware, that we can only test packages for
the current branch, and we have to disable delorean-ci to avoid pulling
in e.g master puppet-tripleo on a newton deployment due to
Depends-On.

Change-Id: I8222da355fa7a7134b695a45d815caa640b8029b