Support heterogeneous OVB environments
This adds support for heterogeneous OVB environments to the
te-broker.  It is primarily intended for scale testing jobs, since
the normal test jobs only deploy a single compute node.  We wouldn't
gain much by using a smaller flavor for that one vm and there is a
cost in complexity to setting up the environment.

Right now this will only work for jobs that deploy just control and
compute nodes.  Support for a third role type for ceph or others
could be added in a similar fashion.

Change-Id: I398d13356b3c15c0c7cd448366186b7589ad93e4