Selectively record heat deploy times to graphite
There are a number of heat resources whose deployment times we
would like to keep track of over time.  Graphite is very good at
that, but because a lot of the heat resources have unique ids
appended to them we can't just capture everything or it overwhelms
Graphite with useless single data points.

To get around that, this patch greps specific resource names out of
the deploy time log file and records only those.  The initial list
is resources that have caused us problems in the past, but we can
add more if we find a need.

Change-Id: I25e7d2a81f6ca7249c0934ad16de8f10c637d825
Related-Bug: 1684272