Don't maintain stable scenario env in upgrade CI
In https://review.openstack.org/#/c/464667 I assumed we wanted to
keep the stable scenario on upgrade, to avoid adding services during
the upgrade process.

It seems we're expecting to run the pingtest for master (or whatever
post-upgrade branch) so were actually depending on the services getting
added during the upgrade.

Change-Id: I83a421a23333a01d56781fb329116ae0036ae48a