Use "manual" image modification method for OVB jobs
There is some yet to be determined issue when using libguestfs to modify
the images that is causing it to take up to 50 minutes. We can not
afford that much overhead in the OVB jobs. Instead, we use the manual
method of converting the image to raw and mounting it to do
modifications.

Closes-Bug: 1694280

Change-Id: I64c0aba5b047ea799991deaf89180f76a636f0fc
Depends-On: Ice4487aa8d263aa7f4a426b7924104a469bd1b52