Allow overriding of multinode common config
The current method of including the multinode common config does
not allow for overriding anything set there. This is because
quickstart.sh will always put "-e" args after the the other config
args.

This patch instead includes it at the top of the multinode playbook.
Anything passed on the command line will always override variables
included this way.

This also moves some variables out of the multinode testenv config
that are not specific to tripleo-ci. This patch depends on a
quickstart patch that puts these variables in the common multinode
config.

Change-Id: Ic4fccd4e71224149e20558af50a8b4f84f0ed67a
Depends-On: I7fd723b5547896693e57c76415ab98b98abfdcfc
Depends-On: I55f705a390ae5b927a9fb0f8d8ec153ea61cc3c9