Use sample environments for OVB configuration
Previously OVB was configured using a monolithic environment file
that had to contain all of the parameters and resource_registry
entries needed for the stack.  Recently support was added for a more
TripleO-like system of sample environment files, which makes the
configuration of OVB much cleaner.  While the old method will be
supported for a period of time, we should go ahead and switch to the
new method because the old one is now deprecated.

The role files unfortunately do not support this sort of configuration
yet, but while working on this change I remembered that roles
inherit their configuration from the parent stack so we don't
actually need to duplicate the network-isolation resource_registry
entries in the role file.  This simplifies the configuration quite a
bit too.

Change-Id: Iab6db834d587081f13c35b5270f15187b0b5fb3b