Switch cistatus page to zuul v3
In the zuul v3 switch, a number of things changed that broke the
cistatus page.  The reviewer name is now Zuul instead of Jenkins,
the job names are legacy-job-name instead of gate-job-name, and not
all of the review comments include a pipeline name.  This change
handles all of those differences and allows us to retrieve job
results again.

Note that it will no longer handle older Jenkins-based jobs.  Given
that it will be all zuulv3 in the future I didn't think it was worth
the effort to support both simultaneously.

Change-Id: I494aff33be6141b2ff82595be75bfe20e6a7e382