Do not rely on openstack-puppet-modules in Queens and beyond
We are removing openstack-puppet-modules package since we now have
individual packages for each Puppet module.
We can still rely on puppet-tripleo module to deploy all dependencies.

Note: this patch will probably break undercloud-container on pike, but
this is fine as we target the feature for Queens, and the job is not
voting anyway.

Change-Id: I86a24516b457c81056942c903d020a5654142dc6
Related-Bug: #1676998