Update testenv-worker for newer novaclient
When we updated the client code on the te-broker to support some
changes that required newer clients we broke the instance liveness
check thanks to a backwards incompatibility in novaclient.  Since
all future te-brokers will be using the newer novaclient this change
just updates the Client creation call to use the new syntax.

Change-Id: Ib3331ee21c533f028007acae355528d9c5edbf76