Fix sub_nodes and sub_nodes_private with multiple nodes
When you use this function with multiple sub nodes, we weren't getting
all the nodes into the /etc/nodepool/sub_nodes and
/etc/nodepool/sub_nodes_private files because the loop walked the values
one by one, and did the rsync in the same loop.

Create the files first in a separate loop so that all the values will be
there before we call rsync.

Change-Id: Id85218da7b8424acd6d1e7eac318d7ab8527dcc0