Use external RHCLOUD variable if set
To use  cloud environment different from RDO cloud it's needed to set this value to something that will
match to the used cloud environment quickstart configuration. It also need to be stored at toci-quickstart config files.

For example, for the RHCLOUD foobar if testing with OVB, the following file have to be installed before running the gate test:

/opt/stack/tripleo-ci/toci-quickstart/config/testenv/ovb-foobar.yml

Change-Id: Ifc88a76df8de32b0a66bc507e337a9b407f17e29