add an update non-voting job to mimimal check
Goal: in two diff commits.
Use the update job to gate:
- tripleo-quickstart
- tripleo-quickstart-extras
- tripleo-upgrade

Use the update job to check non-voting
- tripleo-$other repos

At the moment tripleo-upgrade is ungated.
To properly protect against changes in tripleo
repos it is useful to run the update job as
non-voting to help detect if a tripleo change
caused the update job to fail.

Change-Id: I9b47093fc7fc9397a5058f80d954e730304cf12f