Prepare for DLRN modular build driver structure
Previously, the delorean_setup function modified DLRN's build_rpm.sh
script to make sure that the built rpms would not be installed as
part of the mock build in its chroot.

Since [1], DLRN will change the way it builds packages to use a driver
structure, and the build_rpm.sh script will no longer exist. On
the flip side, the default mock driver gives an option to skip
installation in the mock call, so we can simply use it.

[1] - https://softwarefactory-project.io/r/12039

Change-Id: I278da2ddf15a0b23ab522e63ac58c93043c4816f