Add ability to use a different release per playbook
We are changing the way we are passing releases to the playbooks.
This change adds the ability to use the new release handling script. It
enables it only for 4 jobs. Two should be moved to the new way of
handling releases, two are there to test the backwards compatibility
offered by the script.
If the script is uses and produces an output, the new method will be
activated, by filling a dictionary with playbook as key and release as
value.
If there is no output from the script, the dictionary will be empty and
the default QUICKSTART_RELEASE file will be used instead.

Co-Authored-By: Gabriele Cerami <gcerami@redhat.com>
Co-Authored-By: Ronelle Landy <rlandy@redhat.com>
Change-Id: I6514ba15ff4300ac4bd4fe543d3a4954aeb6f175