Add python script to dynamically compose releases
Currently, we do this in a combination of bash in this
repo and the release files themselves. However, as we begin
to support more different upgrade scenarios, this has proven
to be difficult to maintain.

Instead, we will use this python script to create a dictionary
of releases which describes which release will be used at each
point in the deploy: undercloud_install, undercloud_upgrade,
overcloud_deploy, and overcloud_upgrade.

This commit only provides an intial skeleton to work from.
Follow-up commits will provide:

  1. The ability to get a DLRN hash for each of the above breakpoints
  2. A CLI for the tool so it can be called from the TOCI scripts
  3. The ability to parse a featureset config for upgrade variables
  4. The ability to output the dictionary for consumption by TOCI scripts

Change-Id: Ibf61d7d12230f6714eb7dad91169aa043f5f8417