Add CLI argument parser and YAML file parser
This adds a CLI interface with the following options:
  --stable-release
  --featureset-file
  --output-file
  --log-file

This also adds the code and tests for the YAML file
parsing. I tried adding a specific error case for
when the YAML file fails to parse vs the more generic
case when the file just cannot be opened. However,
mock would not behave for testing the specific case.

We log and raise the exception in both cases though,
and the tests cover both as well.

Change-Id: I0834a0e9b3193c664b377ae7e066fe15239bbfb1