Wire in get_dlrn_hash
This wires in the get_dlrn_hash function to the release dictionary,
so that we end up with an actual dlrn hash instead of a named hash.

Release dictionary tests updated to mock this function, including
verification of the expected calls to get_dlrn_hash.

This also adds more comprehensive logging and unittests to the
get_dlrn_hash function.

Change-Id: I41087bf0c247b933a641ea0da582b5f2b368840e