Add ability to upgrade FROM change under test
Currently, we are only able to upgrade TO a change under test. (n-1->n)
This change adds an argument (--upgrade-from), which allows for testing
that a change does not break upgrading FROM a release. (n->n+1)

The change itself is a no-op other than changing the signature of the
compose_releases_dictionary method.

Change-Id: I141aa5e01a097968a2a3427ebcfb5b8d304c9753