Add new abstract base job and nodeset to zuul configuration
To start the migration to zuulv3, we need a new base job that loses all
the devstack dependencies.
We are basing this new abstract job on the base multinode job. Devstack
is not needed anymore, and all the provided operations we still require
are in the pre-run playbooks.
The new nodesets are compatible with the new base job, and should be
used in the migration instead of the legacy-* ones

Change-Id: I827fddbea3692ebe951f240555e6d37b8807a5dc