Collect /tmp/ansible-mistral-action into CI job logs.
When executing the ansible-playbook mistral
action a directory is created within undercloud's
/tmp/. This directory contains the ansible
configuration files used to run the playbook, as well
as an script which contains the ansible-playbook
command run. All this information can be very
useful for CI job debugging purposes now that
we're moving into an ansible-centric TripleO
operations.

Change-Id: I1ae2d712ce08b705913bc22587984105ce497e64