Fix TAG and playbooks logic for all upgrade type jobs
Some work in:
https://review.openstack.org/#/q/topic:toci_jobtype+(status:open+OR+status:merged)

Broke the containerized undercloud upgrade jobs and probably other
upgrade jobs, the TAG variable was overriden by featureset tags while we
need the upgrade tags as well.
This patch moves the logic after the 'case' that already managed a bunch
of tag. So as a result, we should have tags from the case and the tags
needed for the upgrade jobs.

Some other work in

https://review.openstack.org/582466

Broke our ability to stop at the first playbook failure, so if a
playbook fails, we continue to test the next in the list, and if that
passes, we are reporting a false positive

tripleo-common: Fix overwriting downloaded config files
Depends-On: Ib1d3c68ec3c4048ffc7277daf84834288ea50e48

Change-Id: I8330687ec64f1ebb7e3b100410d5bdc3e67ae7f0
Closes-Bug: #1783399
Closes-Bug: #1783857