Remove tripleo.sh --bootstrap-subnodes add ceph loop device
As part of sprint 16 we want to remove tripleo.sh usage from ci in
particular the invocation of bootstrap-subnodes [1].

This was initially proposed as stand-alone role in [2] but discussion
has led us to keeping only creation of the ceph loop device, added
here as ansible tasks in a tripleo-ci/ceph.yaml. This is included
in the pre-run for the tripleo-ci-base job (zuul-v3) but not for
legacy jobs. That needs to wait for [3] and related testing/other
changes.

This removes any calls to tripleo.sh --bootstrap-subnodes from the
templated .j2 version of toci_gate_test and common_functions and
adds the new tripleo-ci/ceph.yaml in pre-run of tripleo-ci-base

Non templated scripts can be changed in a followup after [3].

[1] https://trello.com/c/GvjcnJB2/850-translate-tripleosh-bootstrap-subnodes-into-a-series-of-tasks
[2] https://review.openstack.org/581026
[3] https://review.rdoproject.org/r/#/c/15000
Depends-On: Ib95b02767488aaddfe95633d03cb118dd492a700
Change-Id: I1616a41914ca7142bd7d936ff906c54249012bbb