Remove tripleo.sh --bootstrap-subnodes from toci_gate_test.sh
In https://review.openstack.org/583195 we are removing the bootstrap
subnodes invocations for the zuul v3 jobs and the templated version
of toci_gate_test.sh.j2. Here we do the same for the legacy bash
version of toci_gate_test.sh and add a deprecation note on the
tripleo.sh --bootstrap-subnodes function itself.

The invocation of tripleo.sh is removed with the now unused functions
echo_vars_to_deploy_env_oooq subnodes_scp_deploy_env

Instead the creation of ceph /dev/loop3 is handled in the base job
pre.yaml added in the Depends-On below

Depends-On: https://review.rdoproject.org/r/15000
Depends-On: Ib95b02767488aaddfe95633d03cb118dd492a700
Change-Id: I039388e4c3d4959270c45e827e76993636bd2990