Soften guard for ceph loop device in ci script so we can move it
As part of [1] we want to drop tripleo.sh --bootstrap-subnodes
from the ci scripts and move to ansible. For zuul v3 jobs we do
that in [2].

For "legacy" and rdo jobs the equivalent change is spread at [3]
and [4]. We can't merge [3] however because it would create the
loop device and so cause the exit being removed in this review
to fail the jobs

Thus we need to soften this guard here and log a warning. Once
this merges first, we can then merge [3] and then [4]

[1] https://trello.com/c/lrSUMaqw/885-translate-tripleosh-bootstrap-subnodes-into-a-series-of-tasks-s17
[2] https://review.openstack.org/#/c/583195/
[3] https://review.rdoproject.org/r/#/c/15000/
[4] https://review.openstack.org/#/c/587012/

Change-Id: Id69c3148bd077e338ef29b9bf8edc57fa2581699