Add a script to compare reviews CI artifacts
The "compare-reviews.py" is a manual execution script that needs a left
hand and right hand review numbers and optionally you can override the
list of files that it's going to compare.

By default the list of files that compare is:
- playbook_executions.log
- reproducer-quickstart.sh
- collect_logs.sh

An example of execution would be "./compare-reviews.py 584508  589068"
It will compare the CI artifacts stated before between the two reviews.

The output is a diff with the differences, for each file at each job.

Change-Id: Ib3c05f327e3d6773990e4e39b1892a61e41cb044