Periodic fs050: add promotion hash in "previous" function
For the first time we are adding an upgrade job to the promotion
pipeline. Since the release in these jobs is
promotion-testing-hash-{release}, we need to add a case to the function
that calculates the previous release of the upgrade.
We'll need to update also the emit_releases_script for this to work with
this featureset.

Change-Id: I89e349b3ff563e5db59173f4751a52783d63f1ac