Reduce unit testing changes required for emit_releases_file
When looking at recent edit of emit_releases_file.py it seems that the
amount of manual unit test changes required [1] is so high that it
might deter contributors and consume too much effort.

Much of the edited unit test data seems repetitive, so my rationale
here is that we can reduce the amount of data resulting in a small
loss of coverage but considerable reduction of required labor when
making emit_releases_file amendments.

The order of side effects for get_hash function was considered for the
test result, now we use a fixture that is backed up by a tuple to a
string dictionary to configure the mock, so the order is not important
anymore.

Co-Authored-By: Quique Llorente <ellorent@redhat.com>

[1] https://review.openstack.org/#/c/606996/7/scripts/emit_releases_file/test_release_name.py

Change-Id: Ie573813a4916aeb1bb15b5aa38dc2a732586e8c2