Fix python3 detection logic
The python_cmd function code is copied from already merged
tripleo-quicstart/quickstart.sh version.

In the future we plan to use it form a single location.

Previous version was using python3 on CentOS-7 if it was installed
which was not desired behavior.

Change-Id: Id03b7ff76054c807e46c2483510313b4de66d678
Story: https://tree.taiga.io/project/tripleo-ci-board/task/153