Switch scenarios to non-voting
Due to recent issues in the gate and the fact that we're using way too
much CI resources, I think it is time to switch the scenario jobs to
non-voting to reduce the amount of load we are using upstream. These
jobs need to be replaced with voting standalone jobs going forward.

See also:
http://lists.openstack.org/pipermail/openstack-dev/2018-October/136141.html
http://lists.openstack.org/pipermail/openstack-dev/2018-October/135396.html

Change-Id: I605a11a900197cc9188156ae74412cc54f1df943