Add new standalone_environment_files into featureset_override
As part of [1] we need the ability to override the environment
file passed to standalone jobs for the different scenarios.
This adds a standalone_environment_files parameter into the
allowable values for featureset_override and it expects a list
of paths, relative to /usr/share/openstack-tripleo-heat-templates.
This list will then be passed to the standalone deployment command
using the standalone_custom_env_files (see [1]). With this we
can avoid needing a new featureset to override this environment
for each of the scenario standalone jobs

[1] https://tree.taiga.io/project/tripleo-ci-board/task/363
Co-Authored-By: Alex Schultz <aschultz@redhat.com>
Change-Id: If36d96a96d98b63881ee096eaae54ffc60b6ecf1