Move ceph-loop3 to ceph-loop-device role for standalone scenarios
This moves ceph /dev/loop3 tasks to a new ceph-loop-device role.
This has no effect in the 'normal' case with tasks included as
always for subnodes.

However we are moving scenario jobs to standalone and so
need these tasks on primary node when we have ceph. A new
parameter on the job definition 'standalone_ceph' will
signal this and include the role accordingly. See [1] for example.
Tracked by ci team at [2]

[1] https://review.openstack.org/#/c/619508/
[2] https://tree.taiga.io/project/tripleo-ci-board/task/405
Change-Id: Icae7d956061e444d3bf9364c7336cd834aa7622a