Create tripleo-ansible-role-job template for ansible projects.
Currently, each ansible project being used in tripleo has
defined its own jobs, or it is being using some python
related template which implies uploading the project to pypi
even though it's an ansible project.
Creating a new job template for the ansible projects allow us
defining the needed jobs to be executed for an ansible project
in a centralized way. In a first iteration only linters are
being included, as some projects do not have docs to build, but
this could be extended in the future.

Change-Id: I153b36aaa3df1821cdaa5d6590a6eaa67eeff5a1