Allow dlrn_hash_tag* vars to be overridden
When running the run-v3 playbook with the reproducer
workflow, a specific dlrn_hash_tag and newest tag
need to be passed to the job. This can be done in
featureset_override - if allowed in the playbook.

Also allows for getting a ready releases.sh file to use
to ensure a static dlrn hash.

Change-Id: I6201cd62e7d314fb8b53a104563c11c28ed8f2a2