[Configuration] Fix tox python3 overrides.
We want to default running all tox environments under python 3, so set
the basepython value in each environment. We do not want to specify a
minor number, because we do not want to have to update the file every
time we upgrade python. We do not want to set the override once in
testenv, because that breaks the more specific versions used in default.

Change-Id: I619708463e67bcdc3527222b1a7ecde32070694a