Avoid insecure umask which breaks loading of ansible.cfg
Ansible had a security fix which now prevents it from loading
ansible.cfg files that are world writable (o+w) which is exactly what
the old jobs did in order to make devstack work with hardlinks across
multiple users.

We replace the hardlinks with CoW and reset the permissions to system
default which is what is to be expected for source code.

Change-Id: Ia1bfc89ee8edecd1ebad3487e8c63bdda4d083a0
Partial-Bug: 1807703