Implement workaround for failure to load ansible.cfg
While the correct fix for this is to stop using fetch-zuul-cloner job,
this is not possible right now as is imported by zuul base job.

infra-root said that they it will take more time for the zuul base
to be fixed due to risk involved on very-old jobs, so we need to
implement a local fix for tripleo.

Change-Id: I14abcb3273e68e267d6bcc7b2a060486bbd77351
Partial-Bug: 1807703